Create  res up
I understand you want a way to make your image 50% larger and save it as a JPG. While I cannot directly manipulate image files myself, I can provide you with an HTML and JavaScript tool that you can use in your own browser to perform this resizing and conversion.

You can copy the code below, save it as an .html file on your computer (e.g., resize.html), and then open that file in your web browser. Once opened, you'll be able to choose your image, and the page will automatically process it to be 50% larger and initiate a download for you as a JPG file.

This solution uses your browser's capabilities to handle the image processing.